requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.42.0
django-bootstrap4==2.2.0
django-countries==6.1.3
Django-CustomUser-Model==1.1.0
django-phone-field==1.8.1
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.4
python3-openid==3.2.0
pytz==2019.3
requests==2.24.0
requests-oauthlib==1.3.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
